feat: Added HtmlStyle
The HtmlStyle class is a styling configuration for the HtmlText composable that will be added in a later commit. This class groups multiple SpanStyles for every by the HtmlText composable supported ParcelableSpan. This makes it possible to use a single parameter in the HtmlText composable to pass all the desired styles.