fix: Fixed possible NPE in RemoteActionCompatParcelizer
In some rare cases setting the custom actions for accessibility would case a NPE in the RemoteActionCompatParcelizer via the SemanticsPropertiesKt.setCustomActions function. Based on crash logs that use this library it seems to be isolated to android 8 devices. Because all the used variables are null-safe, I resolved this by adding an extra check around URLSpan.url (which is java code) and an extra try catch so that this and other throwables are caught and can be logged if desired.