feat: Added UriHandler extension with error handling
The default openUri method of the UriHandler doesn't try to open the URI safely and can throw an exception, i.e. activityNotFoundException. This extension function wraps the opening of the URI in  a simple try catch block, and will invoke the passed onError function if the opening of the URI throws a throwable.